resolve warnings

+++ HW02/matmul.cpp
@@ -1,9 +1,9 @@
 #include "matmul.h"
 
 void mmul1(const double* A, const double* B, double* C, const unsigned int n){
-    for(int i = 0; i < n; i++){
-        for(int j = 0; j < n; j++){
-            for(int k = 0; k < n; k++){
+    for(size_t i = 0; i < n; i++){
+        for(size_t j = 0; j < n; j++){
+            for(size_t k = 0; k < n; k++){
                 C[i*n + j] += A[i*n + k]*B[k*n+j];
             }
         }
@@ -11,9 +11,9 @@ void mmul1(const double* A, const double* B, double* C, const unsigned int n){
 }
 
 void mmul2(const double* A, const double* B, double* C, const unsigned int n){
-    for(int i = 0; i < n; i++){
-        for(int k = 0; k < n; k++){
-            for(int j = 0; j < n; j++){
+    for(size_t i = 0; i < n; i++){
+        for(size_t k = 0; k < n; k++){
+            for(size_t j = 0; j < n; j++){
                 C[i*n + j] += A[i*n + k]*B[k*n+j];
             }
         }
@@ -21,9 +21,9 @@ void mmul2(const double* A, const double* B, double* C, const unsigned int n){
 }
 
 void mmul3(const double* A, const double* B, double* C, const unsigned int n){
-    for(int j = 0; j < n; j++){
-        for(int k = 0; k < n; k++){
-            for(int i = 0; i < n; i++){
+    for(size_t j = 0; j < n; j++){
+        for(size_t k = 0; k < n; k++){
+            for(size_t i = 0; i < n; i++){
                 C[i*n + j] += A[i*n + k]*B[k*n+j];
             }
         }
@@ -31,9 +31,9 @@ void mmul3(const double* A, const double* B, double* C, const unsigned int n){
 
 }
 void mmul4(const std::vector<double>& A, const std::vector<double>& B, double* C, const unsigned int n){
-    for(int i = 0; i < n; i++){
-        for(int j = 0; j < n; j++){
-            for(int k = 0; k < n; k++){
+    for(size_t i = 0; i < n; i++){
+        for(size_t j = 0; j < n; j++){
+            for(size_t k = 0; k < n; k++){
                 C[i*n + j] += A[i*n + k]*B[k*n+j];
             }
         }
